requirements.txt
# Flask and WSGI
Flask
Werkzeug
gunicorn

# PostgreSQL
psycopg2-binary

# Environment variables
python-dotenv

# HMAC and cryptography (for signature validation)
cryptography
byteforge-hmac

# Monitoring and observability
prometheus-client
python-json-logger

# Rate limiting
redis

# Centralized logging with Loki
mazza-base @ git+https://${CR_PAT}@github.com/mazza-vc/python-mazza-base.git@main
